Latest neural net code.
The new layout of MachineLearning/neural_cnn.py is a basic
convolutional neural network experiment.

This model was able to achieve an R^2 score of 0.82 on the
validation set of our latest city round.

Further analysis is needed to verify these results, however.

Architectural changes can be made to improve the performance
of this model, for sure.